chore: rename edit-gas component files; re-export
