+++ components/brave_wallet_ui/components/extension/edit-gas/edit-gas.styles.ts
